examples/google_adk/requirements.txt
google-adk
google-genai==1.16.1
nest-asyncio==1.6.0
agentops
python-dotenv 
deprecated